feat: add ast, token-scanner and token-vec for parser
- Add [Ast] structure model.
- Added [TokenScanner] which is a clone of [ScannerIt] that's made for
  [Peekable<TokenVec>].
- Added [Parser] structure interface with default method
  implementations.
