+++ src/interpreter/ast.rs
@@ -1,0 +1,11 @@
+/*
+ * TODO: add documentation about file.
+ */
+
+use super::{position::Positioned, token::Token};
+
+pub enum Ast {
+    Single(Positioned<Token>),
+    Group(Vec<Positioned<Token>>),
+}
+
